+++ Interface/Components/sidebar.py
@@ -7,6 +7,7 @@ from Managers.chat_history import (load_conversations,
                                    load_specific_conversation,
                                    start_new_conversation)
 from Utils.load_settings import load_data_path
+from Connections.ollama import ollama_version_test
 import sounddevice as sd
 from CTkToolTip import CTkToolTip
 from datetime import datetime, timedelta
@@ -181,6 +182,7 @@ def create_sidebar(globals):
     def reset_to_new_chat():
         """Resets to a new conversation and clears the chat frame."""
         sd.stop()
+        ollama_version_test(globals)
         start_new_conversation(globals)
 
         # Clear the current chat bubbles
@@ -217,6 +219,10 @@ def create_sidebar(globals):
         logging.debug(f"Sidebar toggle called.")
         sidebar.tkraise()
         logging.debug(f"Current sidebar x position: {sidebar.winfo_x()}")
+
+        # Ensure Ollama is running
+        ollama_version_test(globals)
+
         if globals.sidebar_open:
             logging.debug(f"Closing sidebar.")
             slide_out()